// packages/datadocs/src/layout/enums/pane.ts
export enum PaneType {
  PANE = "pane",
  GROUP = "group",
}

export enum PaneSingleType {
  TABBED = "tabbedPane",
  TILED = "tiledPane",
  FIXED = "fixedPane",
  EMBEDDED = "embeddedPane",
}

export enum PaneGroupType {
  HORIZONTAL = "hGroup",
  VERTICAL = "vGroup",
  TABS = "tabsGroup",
  TILED = "tiledGroup",
  FIXED = "fixedGroup",
  EMBEDDED = "embeddedGroup",
  SPREADSHEET = "spreadsheetGroup",
}

export enum Placement {
  ROOT = "root",
  CONTAINER = "container",
  BOTTOM = "bottom",
  CONTAINER_LEFT = "container:left",
  CONTAINER_LEFT_INNER = "container:left:inner",
  CONTAINER_RIGHT = "container:right",
  CONTAINER_RIGHT_INNER = "container:right:inner",
  CONTAINER_CENTER = "container:center",
  CONTAINER_CENTER_MAIN = "container:center:main",
  CONTAINER_CENTER_BOTTOM = "container:center:bottom",
  INNER = ":inner",
  TAB = "tab",
}

export enum PaneName {
  DASHBOARD = "dashboard",
}